Refactor suite parsing code
Accepts "`" and "//`" at the beginning of a test to lead in options for that
specific test (the "//`" is new, allowing to have .c suites with new syntax).
Accepts "`" and "/*`" on their own line to start test output, and "`" and "`*/"
on their own line to end test output (the part about "on their own line" is
new).  Silently convert Unix line endings to DOS line endings.  Drop support
for global test suite options for now (which used to be lead in by "``").